+++ src/Events/PaymentSucceeded.php
@@ -10,5 +10,10 @@ class PaymentSucceeded
 {
     use Dispatchable;
 
-    public function __construct(public readonly PaymentTransaction $transaction) {}
+    public readonly string $gateway;
+
+    public function __construct(public readonly PaymentTransaction $transaction)
+    {
+        $this->gateway = (string) $transaction->gateway;
+    }
 }